Move very idiosyncratic callback to the only place it is needed.
