+++ crystal_diffusion/callbacks/callbacks.py
@@ -1,5 +1,3 @@
-import dataclasses
-
 import matplotlib.pyplot as plt
 import torch
 from pytorch_lightning import Callback
@@ -17,18 +15,6 @@ from crystal_diffusion.score.wrapped_gaussian_score import \
 
 plt.style.use(PLOT_STYLE_PATH)
 TENSORBOARD_FIGSIZE = (0.65 * PLEASANT_FIG_SIZE[0], 0.65 * PLEASANT_FIG_SIZE[1])
-
-
-class HPLoggingCallback(Callback):
-    """This callback is responsible for logging hyperparameters."""
-
-    def on_train_start(self, trainer, pl_module):
-        """Log hyperparameters when training starts."""
-        assert hasattr(
-            pl_module, "hyper_params"
-        ), "The lightning module should have a hyper_params attribute for HP logging."
-        hp_dict = dataclasses.asdict(pl_module.hyper_params)
-        trainer.logger.log_hyperparams(hp_dict)
 
 
 class TensorBoardDebuggingLoggingCallback(Callback):
